requirements.txt
Django==2.2.3
django-filter==2.2.0
djangorestframework==3.10.1
djangorestframework-simplejwt==4.3.0
psycopg2==2.8.3
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
